+++ client/projectmate/src/components/pages/Comment.js
@@ -23,15 +23,17 @@ const Comment =(props) =>{
             setSecret(0);
         }
     }
+
     const clickHandler = (e) =>{
         setInput(!input);
         setSubComment('');
     }
 
+    //답글 보내기
     const postSubComment =async()=>{
         var config = {
             method: `post`,
-            url: `/comment/${props.id}`,
+            url: `/comment/${props.postId}`,
             headers: { 
               'Authorization': `${auth}`,
               'Content-Type': 'application/json'
@@ -72,7 +74,7 @@ const Comment =(props) =>{
             data : {
             "content" : `${modComment}`,
             "parentId" : null,
-            "secret" : `1`
+            "secret" : `${secret}`
             }
           };
           axios(config)
@@ -85,7 +87,7 @@ const Comment =(props) =>{
     }
 
     //삭제
-    const DeleteComment = e =>{
+    const DeleteComment = e =>{ 
         var config={
             method: `delete`,
             url: `/comment/${props.id}`,
@@ -95,7 +97,7 @@ const Comment =(props) =>{
             },
         }
         axios(config)
-        .then(response => console.log('게시글 삭제 성공'))
+        .then(response => console.log('댓글 삭제 성공'))
         .catch(error => {
             console.error(error);
         });
@@ -130,7 +132,7 @@ const Comment =(props) =>{
                     </div>}
                 <div>
                     {props.commentList.map((subcomment)=>(
-                        <SubComment {...subcomment} auth={props.auth} getComments={props.getComments}/>
+                        <SubComment {...subcomment} auth={props.auth} getComments={props.getComments} postId={props.postId}/>
                     ))}
                 </div>
             </div>}
