--- client/projectmate/src/components/pages/components/NonFound.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const NonFound = () => {
-    return (
-        <>
-            <div className='nonfound'><strong>현재 모집중인 프로젝트가 없습니다...</strong></div>
-        </>
-    )
-}
-
-export default NonFound;
