--- client/projectmate/src/components/pages/components/css/Details.css
@@ -1,25 +1,0 @@
-.posting {
-  margin-top: 5%;
-  margin-left:20%;
-  margin-right:20%;
-}
-
-.postingDay {
-  text-align: right;
-  color: gray;
-}
-
-.postingTitle {
-  text-align: left;
-}
-
-.postingInfo {
-  margin-bottom:5%;
-  margin-right:5%;
-  margin-left: 5%;
-}
-
-.ck-editor__editable { 
-  height: 400px; 
-  color: black;
-}
