--- client/projectmate/src/components/pages/components/pages/Division.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Division = (props)=>{
-    return(
-        <div>
-            <button>{props.name}</button>
-        </div>
-)
-}
-
-export default Division;
