--- client/projectmate/src/components/pages/components/pages/NotFound.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-const NotFound = () =>{
-    return(
-        <>
-            <div>
-                <h1 className='not-found'>404 Not Found!!</h1>
-            </div>
-        </>
-    )
-}
-
-export default NotFound;
